+++ src/components/gallery/ImageCarousel.jsx
@@ -1,6 +1,6 @@
 import { IonCol, IonContent, IonGrid, IonIcon, IonRow, IonSlide, IonSlides } from "@ionic/react";
 import { useState } from "react";
-import FlexImage from "../Image/FlexImage";
+import FlexImage from "../image/FlexImage";
 // import { ellipse } from "ionicons/icons";
 import "./styles.css";
 const sampleImage = "https://i1.wp.com/jejuhydrofarms.com/wp-content/uploads/2020/05/blank-profile-picture-973460_1280.png?ssl=1";
