+++ src/components/gallery/styles.css
