+++ src/pages/edit-collection/AddCollection.jsx
@@ -1,4 +1,4 @@
-import CollectionForm from "../../components/Form/CollectionForm";
+import CollectionForm from "../../components/form/CollectionForm";
 
 const EditCollection = (props) => {
   return (
